refactor: use common script to handle ipc creation
