+++ apps/electron/public/game/tools/fast-travels/index.html
@@ -29,8 +29,12 @@
 			<div class="row g-4 mt-1" id="locations"></div>
 		</div>
 
-		<script type="module">
-			require('../../../../dist/renderer/game/scripts/fast-travels.js');
+		<script>
+			require('../../../../dist/renderer/game/scripts/script.common.js');
+
+			window.addEventListener('DOMContentLoaded', () => {
+				require('../../../../dist/renderer/game/scripts/script.fast-travels.js');
+			});
 		</script>
 	</body>
 </html>
