+++ apps/electron/public/game/tools/follower/index.html
@@ -81,8 +81,12 @@
 			</div>
 		</div>
 
-		<script type="module">
-			require('../../../../dist/renderer/game/scripts/script.follower.js');
+		<script>
+			require('../../../../dist/renderer/game/scripts/script.common.js');
+
+			window.addEventListener('DOMContentLoaded', () => {
+				require('../../../../dist/renderer/game/scripts/script.follower.js');
+			});
 		</script>
 	</body>
 </html>
