+++ apps/electron/src/renderer/game/scripts/script.follower.ts
@@ -1,53 +1,4 @@
-import { ipcRenderer } from 'electron/renderer';
-import { WINDOW_IDS } from '../../../common/constants';
 import { IPC_EVENTS } from '../../../common/ipc-events';
-import PortMonitor from '../../../common/port-monitor';
-
-let g_msgPort: MessagePort | null = null;
-
-async function setupHeartbeat() {
-	// New ports are required, if previous ones are closed
-	const channel = new MessageChannel();
-	const transferPort = channel.port1;
-	const msgPort = channel.port2;
-	g_msgPort = msgPort;
-
-	// Start both ports
-	transferPort.start();
-	msgPort.start();
-
-	ipcRenderer.postMessage(IPC_EVENTS.SETUP_IPC, WINDOW_IDS.FOLLOWER, [
-		transferPort,
-	]);
-
-	new PortMonitor(
-		msgPort,
-		() => {
-			console.info('Established ipc with parent.');
-		},
-		() => {
-			msgPort.close();
-			transferPort.close();
-			g_msgPort = null;
-			console.info('Trying to re-establish heartbeat in 1s.');
-			setTimeout(() => {
-				void setupHeartbeat();
-			}, 1_000);
-		},
-		false,
-	);
-
-	msgPort.addEventListener('message', async (ev) => {
-		if (ev.data.type === 'heartbeat' || ev.data.type === 'heartbeat-ack') {
-			return;
-		}
-
-		if (ev.data.event === IPC_EVENTS.FOLLOWER_ME) {
-			(document.querySelector('#player') as HTMLInputElement).value =
-				ev.data.args.name;
-		}
-	});
-}
 
 function toggleState(state: boolean) {
 	{
@@ -114,13 +65,11 @@ function toggleState(state: boolean) {
 	}
 }
 
-window.addEventListener('DOMContentLoaded', async () => {
-	await setupHeartbeat();
-
+window.addEventListener('ready', async () => {
 	{
 		const btn = document.querySelector('#me') as HTMLButtonElement;
 		btn.addEventListener('click', async () => {
-			g_msgPort?.postMessage({ event: IPC_EVENTS.FOLLOWER_ME });
+			window.msgPort?.postMessage({ event: IPC_EVENTS.FOLLOWER_ME });
 		});
 	}
 
@@ -135,7 +84,7 @@ window.addEventListener('DOMContentLoaded', async () => {
 					document.querySelector('#player') as HTMLInputElement
 				).value;
 
-				g_msgPort?.postMessage({
+				window.msgPort?.postMessage({
 					event: IPC_EVENTS.FOLLOWER_START,
 					args: {
 						name,
@@ -167,14 +116,17 @@ window.addEventListener('DOMContentLoaded', async () => {
 					},
 				});
 			} else {
-				g_msgPort?.postMessage({
+				window.msgPort?.postMessage({
 					event: IPC_EVENTS.FOLLOWER_STOP,
 				});
 			}
 		});
 	}
-});
 
-window.addEventListener('beforeunload', () => {
-	g_msgPort?.close();
+	window.addMsgHandler(async (ev) => {
+		if (ev.data.event === IPC_EVENTS.FOLLOWER_ME) {
+			(document.querySelector('#player') as HTMLInputElement).value =
+				ev.data.args.name;
+		}
+	});
 });
